Refactor page object distribution
Created PageObjects folder to contain all the page object classes in order to be accessible and consumed by the different tests
